[feature] Optimize activerecord-import cache invalidation (#19)
### Summary
When using PostgreSQL, we can reduce the number of records we have to invalidate for each import.

Postgres supports "RETURNING id", so we can easily get the result set back and perform invalidation on the created result set.

We also need to send another query before the import in case it's an UPSERT operation. In this case, we'd need to query all the values before updating so we can properly invalidate the cache.

As a result, we are now sending two extra SELECT SQL queries for each import operation. This should not be an issue since it only makes the import slightly slower -- the result set of both queries is strongly correlated with the size of the records to import (cannot be too big).

### Test Plan
- added test cases
- ci